Move secondPillar to flows
