+++ src/entries/components/flows/secondPillar/Steps.scss
@@ -1,4 +1,4 @@
-@import "../variables";
+@import "../../variables";
 
 .px-col {
   padding-left: 15px !important;
