+++ src/entries/components/flows/secondPillar/confirmMandate/ConfirmMandate.scss
