+++ src/entries/components/flows/secondPillar/confirmMandate/fundTransferTable/FundTransferTable.js
