+++ src/entries/components/flows/secondPillar/confirmMandate/fundTransferTable/FundTransferTable.spec.js
