+++ src/entries/components/flows/secondPillar/confirmMandate/fundTransferTable/index.js
