+++ src/entries/components/flows/secondPillar/confirmMandate/index.js
