+++ src/entries/components/flows/secondPillar/confirmMandate/mandateNotFilledAlert/MandateNotFilledAlert.spec.js
