+++ src/entries/components/flows/secondPillar/confirmMandate/mandateNotFilledAlert/index.js
