+++ src/entries/components/flows/secondPillar/selectSources/exactFundSelector/fundExchangeRow/FundExchangeRow.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { PropTypes as Types } from 'prop-types';
 import { withTranslations } from 'retranslate';
 
-import { createClamper } from '../../../../common/utils';
+import { createClamper } from '../../../../../common/utils';
 import './FundExchangeRow.scss';
 
 const clampFromZeroToHundred = createClamper(0, 100);
