+++ src/entries/components/flows/secondPillar/selectSources/exactFundSelector/fundExchangeRow/FundExchangeRow.scss
