+++ src/entries/components/flows/secondPillar/selectSources/pensionFundTable/PensionFundTable.scss
@@ -1,4 +1,4 @@
-@import "../../../variables";
+@import "../../../../variables";
 
 .tv-table__header {
   font-size: 0.8em;
