+++ src/entries/components/flows/secondPillar/selectSources/targetFundSelector/TargetFundSelector.scss
