+++ src/entries/components/flows/secondPillar/stepTitle/StepTitle.scss
@@ -1,4 +1,4 @@
-@import "../../variables";
+@import "../../../variables";
 
 .tv-step__number {
   width: 32px;
