+++ src/entries/components/flows/secondPillar/stepTitle/StepTitle.spec.js
