+++ src/entries/components/flows/secondPillar/success/Success.scss
@@ -1,4 +1,4 @@
-@import "../../variables";
+@import "../../../variables";
 
 .tv-success {
   &__container {
