+++ src/entries/components/flows/secondPillar/success/index.js
