+++ src/entries/components/flows/secondPillar/transferFutureCapital/targetFundTooltipBody/TargetFundTooltipBody.js
