+++ src/entries/components/flows/secondPillar/transferFutureCapital/targetFundTooltipBody/TargetFundTooltipBody.spec.js
