+++ src/entries/components/flows/secondPillar/transferFutureCapital/targetFundTooltipBody/advanced.png
