+++ src/entries/components/flows/secondPillar/transferFutureCapital/targetFundTooltipBody/conservative.png
