+++ src/entries/components/flows/secondPillar/transferFutureCapital/targetFundTooltipBody/index.js
